# openapi.yaml
openapi: 3.0.0
servers:
  - url: https://api.kalastud.io
    description: Production Server
  - url: https://perfect-week-test.herokuapp.com
    description: Test Server
info:
  title: PerfectWeek API
  description: The PerfectWeek web API
  version: 0.1.0
  contact:
    email: perfectweek_2020@labeip.epitech.eu
  license:
    name: MIT
    url: 'https://github.com/PerfectWeek/web-api/blob/dev/LICENSE'

####################
# Tags
#
...
paths:
...
  /calendars:
    post:
      tags:
        - Calendar
      summary: Create a new Calendar
      operationId: createCalendar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CalendarCredentials'
      responses:
        '201':
          description: Calendar created
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Calendar created
                  calendar:
                    $ref: '#/components/schemas/CalendarInfo'
        '400':
          description: Invalid Calendar
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Invalid fields in Calendar
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #
  # Schemas
  #
  schemas:
...
    GroupMemberRole:
      type: string
      enum:
        - admin
        - spectator
        - actor
        - outsider
...
    CalendarCredentials:
      type: object
      properties:
        name:
          type: string
          example: amazing_calendar
    CalendarInfo:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: amazing_calendar
        role:
          $ref: '#/components/schemas/GroupMemberRole'
...
    AuthenticationFailed:
      description: The authentication failed
      type: object
      properties:
        message:
          type: string
          example: You must be authenticated to perform this action
    # 403
...
    InternalServerError:
      description: Internal Server Error
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error